Cartrigbe refactor
